pyproject.toml
[tool.poetry]
name = "project"
version = "2.3.0"
description = "Service for managing projects and project related resources."
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.9"
aiobotocore = "2.3.0"
asyncpg = "0.25.0"
fastapi = "0.75.0"
greenlet = "1.1.2"
opentelemetry-api = "1.11.1"
opentelemetry-exporter-jaeger = "1.11.1"
opentelemetry-instrumentation = "0.30b1"
opentelemetry-instrumentation-fastapi = "0.30b1"
pillow = "9.1.0"
pilot-platform-common = "0.1.0"
psycopg2 = "2.9.3"
pydantic = "1.8.2"
python-magic = "0.4.25"
sqlalchemy = "1.4.32"
uvicorn = { extras = ["standard"], version = "0.17.6" }
email-validator = "1.2.1"

[tool.poetry.dev-dependencies]
alembic = "1.7.7"
faker = "13.3.2"
httpx = "0.23.0"
jq = "1.2.2"
pytest = "7.1.1"
pytest-asyncio = "0.18.1"
pytest-cov = "3.0.0"
pytest-httpx = "0.21.0"
pytest-mock = "3.7.0"
pytest-random-order = "1.0.4"
testcontainers = "3.4.2"

[tool.pytest.ini_options]
addopts = "-v --random-order"
testpaths = "tests"
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
